+++ src/pages/leg/components/button-input-table/button-input-table.component.js
@@ -1,9 +1,15 @@
+var _ = require('underscore');
+
 module.exports = {
     onCreate(input) {
         this.state = {
             compactMode: false,
-            players: input.players
+            players: input.players,
+            current: this.findCurrent(input.players)
         }
+    },
+    findCurrent(players) {
+        return _.find(players, (player) => player.is_current_player);
     },
 
     onButtonPress(score, multiplier, isUndo) {
@@ -12,6 +18,7 @@ module.exports = {
 
     onScoreUpdate(data) {
         this.state.players = data.players;
+        this.state.current = this.findCurrent(this.state.players);
         this.setStateDirty("players");
     }
 }